requirements.txt
djangorestframework
psycopg2-binary
django-cors-headers
